backend/requirements.txt
# Core Dependencies
Flask==3.0.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0

# Testing Dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
requests-mock==1.11.0

# Development Dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1